Add files via upload
Data Analysis of Mortality worldwide

Here in this project, the goal of the project is to collect Mortality data.After collection of data ,we need to prepare the data into correct format and then process the data by filter and cleaning the data.Once this previous step is completed , we need to perform an initial visualizations to formulate a general thesis. Then we start analysis of the data to extract the insights of the data and visualize its using PowerBI to show the results of the analysis.